made updates to themes
