+++ src/pages/ZonePage/ZonesPage.tsx
@@ -25,9 +25,12 @@ import {
 import { Season } from "../../App/models/Season";
 import ErrorBoundary from "../../Components/errorBoundary/ErrorBoundary";
 import { tokens } from "../../theme/theme";
-import { ModalTheme } from "../../theme/ModalThemeInterface";
+import { useModalColorTheme } from "../../theme/ModalTheme";
+import { useMenuBarColorTheme } from "../../theme/MenuBarTheme";
 
 const ZonesPage = () => {
+  const modalColorTheme = useModalColorTheme();
+  const menuBarColorTheme = useMenuBarColorTheme();
   const { season } = useSelector((state: RootState) => state.season);
   const dispatch = useDispatch();
   const [isLoadingZones, setIsLoadingZones] = useState<boolean>(true);
@@ -43,91 +46,11 @@ const ZonesPage = () => {
     };
   };
 
-  const modalColorTheme: ModalTheme = {
-      barButtons: {
-        backgroundColor: colors.menuBar.buttonBackground,
-        color: colors.menuBar.buttonFont,
-        border: "1px solid " + colors.menuBar.buttonBorder,
-        "& .btn-icon": { color: colors.menuBar.buttonIcon + " !important" },
-        "&.action:hover": {
-          border: "1px solid " + colors.menuBar.buttonBorderHover,
-        },
-      },
-      cardModal: {
-        backgroundColor: colors.modal.overlay,
-      },
-      card: {
-        backgroundColor: colors.modal.background,
-        border: "1px solid " + colors.modal.border + " !important",
-        "& .close-icon": {
-          color: colors.modal.closeIcon,
-        },
-        "& .close-icon:hover": {
-          color: colors.modal.closeIconHover,
-        },
-        // Buttons
-        "& .cancel-btn, & .submit-btn, & .img-upload-btn": {
-          border: "2px solid " + colors.modal.buttonBorder,
-        },
-        "& .submit-btn, & .img-upload-btn": {
-          border: "2px solid " + colors.modal.buttonBorder,
-          backgroundColor: colors.modal.buttonBackground,
-        },
-        "& .submit-btn:hover, & .img-upload-btn:hover": {
-          backgroundColor: colors.modal.buttonBackgroundHover,
-          color: colors.modal.buttonFontHover,
-        },
-        "& .cancel-btn": {
-          color: colors.modal.buttonFontHover,
-        },
-        "& .cancel-btn:hover": {
-          backgroundColor: colors.modal.buttonBackground,
-          color: colors.modal.buttonFont,
-        },
-        "& .img-upload-btn, & .submit-btn": {
-          color: colors.modal.buttonFont,
-        },
-        // Fields
-        "& .input-override label, & .img-upload-filename-label, & .dropdown-override label":
-          {
-            color: colors.modal.fieldLabel,
-          },
-        "& .input-override div input, & .input-override.notes .MuiInputBase-multiline textarea, & .img-upload-filename":
-          {
-            color: colors.modal.fieldInputFont,
-          },
-        ".css-hyuuor-MuiButtonBase-root-MuiMenuItem-root, & .dropdown-override div:first-of-type, & .dropdown-unselect::after":
-          {
-            color: colors.modal.fieldInputFont + " !important",
-          },
-        "& .MuiInputBase-formControl, & .MuiInputBase-multiline, & .img-upload-filename, .input-override div input":
-          {
-            backgroundColor: colors.modal.fieldBackground + " !important",
-          },
-        "& .input-override div input:focus, .input-override div:hover input, & .dropdown-override .MuiOutlinedInput-root:hover, .input-override.notes .MuiInputBase-multiline textarea:hover, .input-override.notes .MuiInputBase-multiline textarea:focus":
-          {
-            border: "1px solid " + colors.modal.fieldBorder + " !important",
-          },
-      },
-      cardTitle: {
-        color: colors.modal.titleColor,
-      },
-      cardDescription: {
-        color: colors.modal.description,
-      },
-      gallonsChips: {
-        // borderBottom: "1px solid " + colors.shadow.vary,
-        backgroundColor: colors.whiteBlue.vary,
-        color: colors.gray.toWhite,
-      },
-      gallonsChipsAvatar: {
-        background: colors.whiteBlue.alt,
-        color: colors.primary.const + " !important",
-        "& .bar-gallons-chip-avatar-text": {
-          backgroundColor: colors.whiteBlue.vary,
-        },
-      },
-    };
+  const StyledZoneContainer = `
+    ul[role="listbox"] {
+      background-image: ${colors.menuBar.buttonBackgroundImage} !important; // Dropdown list background
+    }
+  `;
 
   // const fetchSeasons = () => {
   //   agent.Seasons.list().then((seasons) => {
@@ -204,11 +127,13 @@ const ZonesPage = () => {
   return (
     <>
       <ErrorBoundary fallback="Unable to retrieve data for zones. The server may be down.">
+        <style>{StyledZoneContainer}</style>
         <ZoneBar
           fetchZones={fetchZones}
           updateLocalStorageSeason={updateLocalStorageSeason}
           isLoadingZones={isLoadingZones}
           modalColorTheme={modalColorTheme}
+          menuBarColorTheme={menuBarColorTheme}
         />
         <Grid id="zone-grid-background" sx={zonePageColorTheme().zoneGrid}>
           <ZoneList
